#177 Fix digest dispatcher's admin settings and its incorrect behavior
Digest dispatcher (i.e. cron job) has a menu in the Admin Settings
that was displaying time of day one hour ahead of the actual internal
value (e.g. 0 -> 1:00, 3600 -> 2:00, etc.). This has been fixed. There
was also an override that skipped the time check that was left in the
code - this has been removed in order to enable its normal function.