#186 Fix bugged value passing behavior of teachervisible checkbox
The element type or the "teachervisible" field used to be
"advcheckbox", which caused inconsistent behavior after being frozen
i.e. the default value was always passed instead. The element type
has been changed to "checkbox". The element is now properly frozen in
the slot review form and its value is passed properly.