#333 Fix "show past slots" behavior in student view for all users
The error occured because the js code tried to access a non-existing
checkbox element "show my slots only" in the student view, causing
the code to stop executing before the event listener could be
attached to the "show past slots" checkbox. This has been fixed by
adding the "show my slots only" in student view as well, but hiding
it to avoid null reference issues.