#307 Make print preview table scrollable in Firefox
This bug is caused by the apparent incompatibility of div and
fieldset elements when making a scrollable div box. The problem
was exacerbated by QuickForm, which forced all content defined on the
form within fieldsets, either visible or hidden. This has now been
fixed at the cost of losing the visible label of the print preview
table. Further CSS formatting might be required to achieve a more
desirable appearance.