#311 Fix permission checking function for registration-related actions
This function (actually, two: one for displaying buttons and the other
for checking permission after submitting - this might be changed in
future patches, but it works OK now)included an outdated check of
student attendance before the "allow reappointment" functionality was
added. This created a problem when a student didn't attend and wasn't
allowed a reappointment.