Fix wrong use of form data with time unit
The time unit is separately available in the form submission values
accessible in function ‘definition_after_data’ (and passed to
‘_validation_step1’). However, in the parameter ‘$data’ which is passed
to function ‘validation’, the time units have been applied in a
calculation, and the values are in seconds.

Ensure that we do not fail due to this. The issue can be triggered from
the time slots adding form if the location is missing, but a non-zero
gap is entered.