Properly handle reported error condition
In the form used for adding slots, when a few slots are prepared, e.g. a
slot start weekday is changed back to “Choose”, and then e.g. the
duration is changed, this leads to a JavaScript error. Reason is that
even though the error is detected by ‘getslots’ on a weekday change
event, it is not reacted upon properly by the caller.

This commit fixes the error handling: if an error condition is signalled
by function ‘getslots’ in the JavaScript code, the corresponding row
forecast is simply cleared.

This is not perfect (the _source_ of the error could be made clear to
the user), but replaces the current situation of failing with a
well-defined fallback that users may understand.