Harden against wrong (e.g. negative) duration/gap
The form for adding new time slots has fields for duration and for gap.
Both expect an integer value and can be freely edited by users. If users
enter a negative duration or gap value, this can result in an endless
JavaScript loop. Reason is that the negative value is not expected, but
is used in the abort condition of a loop. This may result in a lot of
wasted cycles and a bad user experience.

This commit improves things by hardening the code which works with the
entered integer numbers. All uses of the values get additional checks.
Invalid values result in clearing the calculations for the slot
forecasts in the rows.

While this solution may not be perfect yet, it protects against annoying
(client-side) problems and improves the user experience.

Looking at locallib.php, zero durations are forbidden (which is rather
obvious), and a negative gap (overlapping start dates) is not supported
as well. Both are taken into account in this commit.